decopule AbstractArrayFixer
